Create README.md
How it works?
The tool generates a malicious HTTPS page using Serveo or Ngrok Port Forwarding methods, 
and a javascript code to cam requests using MediaDevices.getUserMedia.

The MediaDevices.getUserMedia() method prompts the user for permission to use a media input which produces a MediaStream with tracks containing the requested types of media. 
That stream can include, for example, 
a video track (produced by either a hardware or virtual video source such as a camera, 
video recording device, screen sharing service, and so forth), 
an audio track (similarly, produced by a physical or virtual audio source like a microphone, 
A/D converter, or the like),
and possibly other track types.